Transition GET /api/v1/chatmessage/:id
